Fix echo -n.
This is a conceptual fix since it can't manifest itself for the set of
variables and values that it currently involves. Using echo -n when the
new value of variable X is empty incorrectly leaves the previous value
of X set in the cache instead of clearing it. So I eliminated -n from
all affected echos for good measure, even though the connected variables
aren't empty, usually. Note that nearby INITSEARCH can be empty, so
there leaving -n would really hurt.