Filmstrip viewer: improve layout & make resizable.
However, 'resizable' doesn't mean that images will scale: they won't.
With this new layout an image fits 100% horizontally and overflows
vertically. The excess vertical image, if any, is initially hidden
and can be revealed by resizing the viewer window vertically.
This layout is pixel-perfect and the basic styles in gtkrc-2.0 work as
expected; they change the background color of the named widgets.