Redo PageUp/Down hotkeys as menu accelerators.
Issue #12 says that findnrun 1.10.6 is snappier that 2.1.1 when typing
search terms - and it is. One possible reason is that the varENTRY
gtkdialog code block includes too many key-press-event handlers, which
could slow down typing. FSR suggested to replace key-press-events with
menu key accelerators. This commit re-implements the PageUp and PageDown
key-press-event handlers as menu accelerators.  Note: The menubar cannot
be completely hidden, and (depending on what GTK theme is in use) a tiny
1x1 pixel artifact could be visible.