Fix duplicated history entries. Close #11.
put-1
Fix a problem with duplicated history entries. The currently selected
search result item was incorrectly inserted in the history list when
the user deleted a history item.

**Usage Notes for the History List**

The top item in the history pull-down list corresponds to the currently
selected item of the search result list, and it is not part of the
history itself. The top item is there for you to grab it, by pressing
the down arrow key, and edit it before activating it, if you want. Note
that since the top item is not part of the history, you just cannot
delete it. You can delete all the following items. But the top item
you cannot delete - it will always come back under the guise of the
currently selected search result list item.

**Power-user Tip** Redefine hotkey F2 as `HOTKEY_F2=1:S-Down:0xff54`.
This gives you `[Shift-Down]` arrow to cycle between the search input
field and the command entry / history list field. `[Down]` alone hops
from the search input field to the search result list; there `[Down]`
selects through the list; now press `[Shift-Down]` to move to the
history pull-down entry field; finally press `[Down]` to recall the
selected search result item and edit/activate it.
pub-0