New sample plugin: fuzzy application search.
Plugin id: agrep_apps
pub-1
New plugin: approximate (fuzzy) application search. This plugin looks
and works like the built-in application search, but it performs
approximate searches that tolerate typing errors to some degree. You
need to install the TRE agrep command for this plugin to work. 32-
and 64-bit pet and txz packages are linked for download in the
[2.2.0 release](https://github.com/step-/find-n-run/releases/2.2.0)
page.
pub-0