...'save results' - for source filmstrip - done.
pub-1
New: You can save search results by pressing the new hotkey F4.
If `xclip` is installed, tab-delimited data are copied to the clipboard,
otherwise they are sent to standard error strem - so you might want to
run findnrun in a terminal to see the data.

You can specify an alternative clipboard command, or send data to
another filter and/or file. This is described under XCLIP in the
_Setting Preferences_ of the main help file.

The ability to save search results extends to all built-in sources and
also to the example plugins: Filmstrip, and Approximate Application Search.

Plugin developers can easily enable the save function for their plugin,
cf. the plugin development help documentation.
pub-0