Replace generate_about_text() -> set_ABOUT_TEXT (note)
Motivation: eliminate one sub-shell invocation to speed up program
start. Of course, it shaves just a few milliseconds but every penny
counts the design goal "start as fast as possible".

Essentially this change replaces $() of several "echo -n" lines with a
global variable initialized with a line continuation of several strings.
This change leaves quoting unchanged so the ease of reading the code is
unchanged. I'm going to apply this type of change to all the generate_*
functions.