New: icons honor the GTK2 icon theme, if any (note)
New preference in ~/.findnrunrc: `GTK_ICON_THEME_NAME` may be set to the
name of an installed icon theme. This value is automatically initialized
from ~/.gtkrc-2.0 so in most cases you shouldn't worry. Note that if you
use GTK3, this value will be initialized from ~/.gtkrc-2.0 and you may
want to edit it in ~/.findnrunrc.