gtkdialog set multiple vars with one command (note)
Before this change each variable in {TAP, DRAIN, ... } was refreshed
by sourcing the SOURCE's declaration file. So this was a large number of
shell invocations. Now each variable is set with <file input> from
a file. The set of files is created with a single awk command. It is
refreshed when the SOURCE changes by effect of hotkeys F3, Ctrl+N, etc.