Change: revamped Saving Search Results F4 (note)
The previous implementation was sorely lacking. Now F4 saves
a a timestamped TSV file in $HOME with a confirmation message.

Variables XCLIP (preferences.md) and FNRSAVEFLT (plugin-dev.md)
are no longer supported. A plugin can use the built-in save
filter, which can be customized by setting variables RDR and CUT
in the shell environment or in ~/.filmstrip or - for individual
sources - in the SAVEFLT_<source-id> definition.
The built-in save filter supports saving to files and to command
pipes, including the keyboard e.g. with RDR="|xclip" if xclip is
installed.